refactor(m5): revise extraction — detect-preprocess to reco-detect, delete reco-gpu-interop
Reverses the M5 extraction (commits 5df39db, 42d54af, 85488ca) and
replaces it with the right refactor, prompted by user observation
that the split the plan imagined doesn't match the actual
dependency graph.

## Why the original extraction was wrong

After moving 4 of 8 files (CUDA trio + CoreML) and hitting the
GpuContext-reshape wall on the remaining 4, two things became
clear:

1. **"Platform interop" isn't one category, it's two.**
   - Truly-external API wrappers (CUDA driver, CoreML): separable.
   - wgpu-native platform paths (Metal texture cache, Vulkan
     external memory): inseparable from the reco-core pipeline
     because they ARE how reco-core talks to wgpu on each
     platform. Extracting them needs either `GpuContext` to
     leave reco-core (30-file ripple) or every interop signature
     to take raw wgpu primitives (invasive).

2. **Several "interop" files are actually reco-detect's concern.**
   - `cuda_kernels` (normalize + HWC→CHW for YOLO input): only
     reco-detect calls it.
   - `npp_interop` (NV12→RGB color conv): only reco-detect.
   - `coreml_inference` (CoreML model wrap): only reco-detect.
   - `metal_compute` (Metal compute for YOLO preprocess): only
     reco-detect.
   They lived in reco-core historically but match reco-detect's
   domain perfectly — detection-input preprocess, not GPU pipeline
   infrastructure.

Gyroflow's architecture (the obvious reference for multi-GPU-
backend video processing) keeps platform interop co-located with
its pipeline rather than in a separate crate. Our use case is the
same plus a detection concern; the detection concern has a natural
home that isn't reco-core.

## GpuEncoder (M7) sanity check

Confirmed this refactor does not block M7:
  - NVENC needs `cuda_interop` (stays in reco-core) ✓
  - VideoToolbox needs `metal_interop` (stays in reco-core) ✓
  - M7 code lives in reco-io (alongside FFmpeg encoder); imports
    the wgpu-native paths from reco-core just like everyone else.
  - No cross-crate inversion, no circular deps — the "prepares
    for K7" motivation in the plan was anticipatory, not
    structural.

## Changes

  - **Deleted** `crates/reco-gpu-interop/` entirely (was scaffold +
    CUDA trio + CoreML from phase 1 + 1.5).
  - **Moved to reco-core** (back): `cuda_interop.rs` — genuine
    shared interop used by zero-copy bridge + detect.
  - **Moved to reco-detect** (new home): `cuda_kernels.rs`,
    `npp_interop.rs`, `coreml_inference.rs`, `metal_compute.rs`
    (the last one was previously in reco-core, also migrates).
  - **Stays in reco-core**: `metal_interop.rs`, `vulkan_interop.rs`,
    `zero_copy.rs` — wgpu-native platform paths for the stitch
    pipeline's zero-copy bridge.

## Cargo.toml shuffles

  - Workspace: `reco-gpu-interop` removed from members.
  - `reco-core/Cargo.toml`: removed the two `reco-gpu-interop`
    target-cfg dependency lines; the Apple frameworks stay because
    `metal_interop` still uses `metal` + `foreign-types`.
  - `reco-detect/Cargo.toml`: gained `libloading` (for `npp_interop`
    dynamic lib load), `block2`, `foreign-types`, `metal`,
    `objc2`, `objc2-foundation`, `objc2-metal`, `objc2-core-ml`
    (all macOS-cfg) for the migrated files.

## Consumer impact

Compile-time renames inside reco-detect only:

  - `use reco_core::cuda_kernels::*` → `use crate::cuda_kernels::*`
  - `use reco_core::npp_interop::*` → `use crate::npp_interop::*`
  - `use reco_core::coreml_inference::*` → `use crate::coreml_inference::*`
  - `use reco_core::metal_compute::*` → `use crate::metal_compute::*`

`use reco_core::cuda_interop::*` paths stay unchanged everywhere.
No external consumer (reco-io, reco-cli, reco-gui, reco-obs,
reco-autocam) needed any edit.

## Verified locally

  cargo check --workspace                                  ✓
  cargo clippy --workspace --all-targets -- -D warnings    ✓
  cargo fmt --all -- --check                               ✓
  cargo test --workspace --lib --features reco-io/config
      -- --test-threads=1                                  ✓ (268 tests)

## Plan status

M5 is closed as "revised / subsumed into the right split" rather
than as originally written:

  | Plan                                      | Status           |
  |-------------------------------------------|------------------|
  | M1 hotfixes                               | done (pre-session) |
  | M2 CI + all-features matrix               | done (pre-session) |
  | M3 StitchCore refactor (9 steps)          | done             |
  | M4 Cross-consumer extractions (4 items)   | done             |
  | M5 reco-gpu-interop extraction            | **revised** — new crate abandoned; preprocess moved to reco-detect, wgpu-native paths stay in reco-core |
  | M6 Live production (K4)                   | pending          |
  | M6.5 Stacked video + reco-control         | pending          |
  | M7 GpuEncoder + perf squeeze              | pending (unblocked - see above) |
  | M8 v2-public-ready polish                 | pending          |

Plan doc update to record the deviation goes in a follow-up
vault-side commit.
