+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "crates/reco-autocam",
     "crates/reco-calibrate",
     "crates/reco-detect",
-    "crates/reco-gpu-interop",
     "crates/reco-gui",
     "crates/reco-obs",
 ]
