+++ crates/reco-detect/Cargo.toml
@@ -24,7 +24,36 @@ reco-core = { path = "../reco-core" }
 ort = { version = "2.0.0-rc.12", optional = true }
 dirs = { version = "6", optional = true }
 log = "0.4"
+thiserror = "2"
 tracing = { workspace = true, optional = true }
+
+# libloading: used by npp_interop (dynamically loads NPP libraries so
+# the compiled binary keeps running on machines without CUDA installed,
+# reporting NotAvailable instead of failing to start).
+[target.'cfg(any(target_os = "linux", target_os = "windows"))'.dependencies]
+libloading = "0.9"
+
+# Apple frameworks: used by coreml_inference (CoreML model wrap) and
+# metal_compute (Metal compute pipeline for YOLO preprocess). These
+# are detect-preprocess concerns that used to live in reco-core; they
+# moved here once it was clear they're only consumed by reco-detect's
+# backends (see M5 revised analysis).
+[target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
+block2 = "0.6"
+foreign-types = "0.5"
+metal = "0.33"
+objc2 = "0.6"
+objc2-foundation = { version = "0.3", features = ["NSObject"] }
+objc2-metal = "0.3"
+objc2-core-ml = { version = "0.3", features = [
+    "MLModel",
+    "MLMultiArray",
+    "MLFeatureValue",
+    "MLFeatureProvider",
+    "MLDictionaryFeatureProvider",
+    "MLModelConfiguration",
+    "block2",
+] }
 
 [build-dependencies]
 cc = "1"
