+++ crates/reco-detect/src/cuda_kernels.rs
@@ -8,7 +8,7 @@
 use std::ffi::c_void;
 use std::sync::OnceLock;
 
-use crate::cuda_interop::{CUdeviceptr, CudaInteropError, CudaKernel};
+use reco_core::cuda_interop::{CUdeviceptr, CudaInteropError, CudaKernel};
 
 /// PTX source for the P010-to-NV12 conversion kernel.
 ///
@@ -103,11 +103,11 @@ fn get_p010_kernel() -> Result<&'static CudaKernel, CudaInteropError> {
 ///
 /// Note: this performs a 2D-pitched to linear conversion. If the source
 /// has padding (pitch > width), use [`p010_plane_to_nv12`] instead, which
-/// handles pitched layouts via [`cuda_2d_copy`](crate::cuda_interop::cuda_2d_copy)
+/// handles pitched layouts via [`cuda_2d_copy`](reco_core::cuda_interop::cuda_2d_copy)
 /// style semantics. For contiguous data (pitch == width * 2), this function
 /// is simpler.
 pub fn p010_to_nv12(src: CUdeviceptr, dst: CUdeviceptr, n: u32) -> Result<(), CudaInteropError> {
-    crate::cuda_interop::cuda_ensure_context()?;
+    reco_core::cuda_interop::cuda_ensure_context()?;
     let kernel = get_p010_kernel()?;
 
     let block = 256u32;
@@ -128,7 +128,7 @@ pub fn p010_to_nv12(src: CUdeviceptr, dst: CUdeviceptr, n: u32) -> Result<(), Cu
     }
 
     // Synchronize to ensure kernel completion before using the output.
-    crate::cuda_interop::cuda_synchronize()?;
+    reco_core::cuda_interop::cuda_synchronize()?;
 
     Ok(())
 }
@@ -308,7 +308,7 @@ pub fn normalize_hwc_to_chw(
     width: u32,
     height: u32,
 ) -> Result<(), CudaInteropError> {
-    crate::cuda_interop::cuda_ensure_context()?;
+    reco_core::cuda_interop::cuda_ensure_context()?;
     let kernel = get_kernel()?;
 
     let block = (16u32, 16u32, 1u32);
@@ -331,7 +331,7 @@ pub fn normalize_hwc_to_chw(
     }
 
     // Synchronize to ensure kernel completion before using the output.
-    crate::cuda_interop::cuda_synchronize()?;
+    reco_core::cuda_interop::cuda_synchronize()?;
 
     Ok(())
 }
