+++ crates/reco-detect/src/detectors/metal.rs
@@ -14,10 +14,10 @@
 
 use std::path::Path;
 
-use reco_core::coreml_inference::CoreMlModel;
+use crate::coreml_inference::CoreMlModel;
+use crate::metal_compute::MetalPreprocessPipeline;
 use reco_core::detector::{CameraId, Detection, DetectorError, DetectorFrame, UnifiedDetector};
 use reco_core::gpu::GpuContext;
-use reco_core::metal_compute::MetalPreprocessPipeline;
 use reco_core::metal_interop::CVPixelBufferRef;
 
 use super::postprocess;
