+++ crates/reco-detect/src/detectors/ort_gpu.rs
@@ -8,17 +8,17 @@
 use std::ffi::c_void;
 use std::path::Path;
 
+use crate::cuda_kernels::normalize_hwc_to_chw;
+use crate::npp_interop::{NppiRect, npp_mirror_c3, npp_nv12_to_rgb, npp_resize_c3};
 use ort::memory::{AllocationDevice, AllocatorType, MemoryInfo, MemoryType};
 use ort::session::Session;
 use ort::value::{Shape, TensorRefMut};
 use reco_core::cuda_interop::{
     CUdeviceptr, cuda_ensure_context, cuda_mem_alloc, cuda_mem_free, cuda_memset_d8,
 };
-use reco_core::cuda_kernels::normalize_hwc_to_chw;
 use reco_core::detector::{
     CameraId, Detection, DetectorError, DetectorFrame, GpuNv12Frame, UnifiedDetector,
 };
-use reco_core::npp_interop::{NppiRect, npp_mirror_c3, npp_nv12_to_rgb, npp_resize_c3};
 
 use super::postprocess;
 
@@ -74,7 +74,7 @@ impl OrtGpuDetector {
         labels: Vec<String>,
         is_10bit: bool,
     ) -> Result<Option<Self>, Box<dyn std::error::Error>> {
-        if !reco_core::npp_interop::is_npp_available() {
+        if !crate::npp_interop::is_npp_available() {
             log::warn!("OrtGpuDetector: NPP not available, GPU detection disabled");
             return Ok(None);
         }
@@ -232,7 +232,7 @@ impl OrtGpuDetector {
                 ));
             }
             // Convert Y plane: width * height samples.
-            reco_core::cuda_kernels::p010_plane_to_nv12(
+            crate::cuda_kernels::p010_plane_to_nv12(
                 y_ptr,
                 y_pitch,
                 self.nv12_8bit_y,
@@ -242,7 +242,7 @@ impl OrtGpuDetector {
             .map_err(|e| DetectorError::InferenceFailed(format!("P010->NV12 Y conversion: {e}")))?;
             // Convert UV plane: width * (height/2) samples.
             // UV plane has width/2 pixel pairs, each 2 u16 values = width u16 samples per row.
-            reco_core::cuda_kernels::p010_plane_to_nv12(
+            crate::cuda_kernels::p010_plane_to_nv12(
                 uv_ptr,
                 uv_pitch,
                 self.nv12_8bit_uv,
