+++ crates/reco-detect/src/metal_compute.rs
