--- crates/reco-gpu-interop/Cargo.toml
@@ -1,85 +1,0 @@
-[package]
-name = "reco-gpu-interop"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-description = "Platform GPU interop glue for Reco (CUDA, NPP, Metal, Vulkan, Android-AHB, iOS-IOSurface)"
-
-# ---------------------------------------------------------------------------
-# Features
-# ---------------------------------------------------------------------------
-#
-# Default features are deliberately empty: headless / test consumers
-# that only want the reco-core pipeline don't pay for CUDA or Metal
-# deps. Consumers opt in via feature flags, typically through a
-# platform-cfg table in their own Cargo.toml:
-#
-#   [target.'cfg(any(target_os = "linux", target_os = "windows"))'.dependencies]
-#   reco-gpu-interop = { path = "...", features = ["cuda"] }
-#
-#   [target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
-#   reco-gpu-interop = { path = "...", features = ["metal"] }
-#
-# `android-ahb` and `ios-iosurface` are stub-only today; they exist so
-# future mobile work has an obvious home and the feature-matrix CI
-# exercises the stub compile.
-[features]
-default = []
-
-# CUDA family: CUDA driver FFI, custom kernels, NPP color conversion.
-# Enables `cuda_interop`, `cuda_kernels`, `npp_interop`.
-cuda = []
-
-# Metal / CoreML family. Stub today; `metal_interop`, `metal_compute`,
-# `coreml_inference` migrations land in a follow-up (they take
-# `&GpuContext` today and need a small reshape to take wgpu device/
-# queue directly).
-metal = []
-
-# Vulkan external-memory extensions. Stub today; `vulkan_interop`
-# migration lands in a follow-up alongside the zero-copy plumbing.
-vulkan-ext = []
-
-# Android hardware buffer (AHardwareBuffer). Stub; mobile readiness
-# placeholder.
-android-ahb = []
-
-# iOS IOSurface. Stub; mobile readiness placeholder.
-ios-iosurface = []
-
-# ---------------------------------------------------------------------------
-# Dependencies
-# ---------------------------------------------------------------------------
-[dependencies]
-log = "0.4"
-thiserror = "2"
-
-# `cuda`-feature deps. Loaded dynamically via libloading so
-# CUDA/NPP libraries are optional at runtime (reco-gpu-interop
-# compiled with `cuda` still runs on a machine without CUDA
-# installed; the relevant functions return `NotAvailable`).
-[target.'cfg(any(target_os = "linux", target_os = "windows"))'.dependencies]
-libloading = "0.9"
-
-# `metal`-feature deps: objc2-family for CoreML + Metal FFI via
-# Apple's objc2 ecosystem. Only compiled on macOS / iOS targets.
-# `metal_interop` and `metal_compute` migrations land in a
-# follow-up (they take `&GpuContext` today); `coreml_inference`
-# is standalone and moves with this commit.
-[target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
-block2 = "0.6"
-objc2 = "0.6"
-objc2-foundation = { version = "0.3", features = ["NSObject"] }
-objc2-core-ml = { version = "0.3", features = [
-    "MLModel",
-    "MLMultiArray",
-    "MLFeatureValue",
-    "MLFeatureProvider",
-    "MLDictionaryFeatureProvider",
-    "MLModelConfiguration",
-    "block2",
-] }
-
-[dev-dependencies]
